embedding/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "embedding"
version = "0.1.0"
description = "Generate embeddings for text segments"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.60.0",
    "openpyxl>=3.0.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
embedding = "embedding.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["embedding*"]
